refactor: encapsulate opened BLOB file handling
Rename blob_send_utils to opened_blob_file and move the send-side
open/read/close handling into a move-only RAII class. This keeps FILE*
ownership private while preserving the existing TCP/RDMA BLOB send behavior.
